+++ src/main/java/lotto/global/constants/MessageType.java
@@ -1,4 +1,4 @@
-package lotto.global.constans;
+package lotto.global.constants;
 
 public enum MessageType {
     COST_REQUEST_MESSAGE("구입금액을 입력해 주세요."),
