+++ src/test/java/lotto/domain/BonusTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import static lotto.global.constans.ErrorMessage.BONUS_NUMBER_DUPLICATE_ERROR;
+import static lotto.global.constants.ErrorMessage.BONUS_NUMBER_DUPLICATE_ERROR;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
